requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
coverage==7.8.0
Django==5.2
django-filter==25.1
djangorestframework==3.16.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
packaging==24.2
pillow==11.2.1
psycopg2==2.9.10
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
stripe==12.0.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0